yeaphone: Applied Davor's patch yeaphone-p1k-callerid-003.patch
The patch was applied with small modifications, i.e. the code can now 
deal with an unpatched version of liblinphone 2.1.1 where 
GSTATE_CALL_IN_INVITE does not come with a caller ID and where after 
each call GSTATE_REG_FAILED is reported.
However the workarounds are only performed if the buggy linphone version 
is detected!